v4

+++ t.c
@@ -2,12 +2,13 @@
 #include <stdlib.h>
 #include <string.h>
 #include <stdbool.h>
+#include <stdint.h>
 
 int count;
 struct Rule
 {
-    int net[4];
-    int netmask[4];
+    uint8_t net[4];
+    uint8_t netmask[4];
     int vlan;
 } list;
 
@@ -18,7 +19,7 @@ void read_config(struct Rule *s, char *filename, int n)
     fscanf(fin,"%d\n",&n);
     count = n;
     for (int i = 0; i < n; i++)
-	fscanf(fin, "net %d.%d.%d.%d netmask %d.%d.%d.%d vlan %d\n", &s[i].net[0], &s[i].net[1],  &s[i].net[2], &s[i].net[3], &s[i].netmask[0], &s[i].netmask[1], &s[i].netmask[2], &s[i].netmask[3], &s[i].vlan);
+	fscanf(fin, "net %hhu.%hhu.%hhu.%hhu netmask %hhu.%hhu.%hhu.%hhu vlan %d\n", &s[i].net[0], &s[i].net[1],  &s[i].net[2], &s[i].net[3], &s[i].netmask[0], &s[i].netmask[1], &s[i].netmask[2], &s[i].netmask[3], &s[i].vlan);
 }
 
 int main()
@@ -28,5 +29,5 @@ int main()
     int n;
     read_config(rule, iface_config, n);
     for (int i = 0; i < count; i++)
- 	printf("%d %d %d %d - %d %d %d %d - %d\n", rule[i].net[0], rule[i].net[1], rule[i].net[2], rule[i].net[3], rule[i].netmask[0], rule[i].netmask[1], rule[i].netmask[2], rule[i].netmask[3], rule[i].vlan);
+ 	printf("%u %u %u %u - %u %u %u %u - %d\n", rule[i].net[0], rule[i].net[1], rule[i].net[2], rule[i].net[3], rule[i].netmask[0], rule[i].netmask[1], rule[i].netmask[2], rule[i].netmask[3], rule[i].vlan);
 }
